\name{arrayAve-methods}
\docType{methods}
\alias{arrayAve}
\alias{arrayAve,pepArrayPP-method}
\title{Methods for function arrayAve}
\description{
Calculates the mean of intra-array replicate features.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "pepArrayPP")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
