% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arrayBatchCorrect-methods.R
\docType{methods}
\name{arrayBatchCorrect}
\alias{arrayBatchCorrect}
\alias{arrayBatchCorrect,ExpressionSet-method}
\alias{arrayBatchCorrect,MultiSet-method}
\alias{arrayBatchCorrect,matrix-method}
\title{Batch Correction for peptide array data (using ComBat algorithm)}
\usage{
arrayBatchCorrect(x, ...)

\S4method{arrayBatchCorrect}{MultiSet}(x, batch, mod, ...)

\S4method{arrayBatchCorrect}{ExpressionSet}(x, batch, mod, ...)

\S4method{arrayBatchCorrect}{matrix}(x, batch, mod, ...)
}
\arguments{
\item{x}{MultiSet, ExpressionSet or matrix object}

\item{batch}{vector indicating batch}

\item{mod}{model matrix for outcome of interest other than the batch}
}
\value{
MultiSet, ExpressionSet or matrix object with batch corrected intensities
}
\description{
Wrapper for the ComBat function within the sva package
}

