\name{arrayBGcorr-methods}
\docType{methods}
\alias{arrayBGcorr}
\alias{arrayBGcorr,pepArrayPP-method}
\title{Methods for Function arrayBGcorr}
\description{
Applies background correction algorithms to array data.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "pepArrayPP")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
