% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/arrayControlMA-methods.R
\docType{methods}
\name{arrayControlMA}
\alias{arrayControlMA}
\alias{arrayControlMA,ExpressionSet-method}
\title{Create M and A values for each peptide array using a negative control array}
\usage{
arrayControlMA(x, ...)

\S4method{arrayControlMA}{ExpressionSet}(x, control.array)
}
\arguments{
\item{x}{ExpressionSet object}

\item{control.array}{Integer value corresponding to the column index of the negative control array}
}
\value{
ExpressionSet object with M and A matrices added to the assayData environment
}
\description{
Calculates M-values (log2 signal ratios) and A-values (log2 average signal) by comparison of
each peptide array with a negative control array.
}

