% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{bg}
\alias{bg}
\alias{bg,MultiSet-method}
\title{Background Intensity Accessor}
\usage{
bg(x)

\S4method{bg}{MultiSet}(x)
}
\arguments{
\item{x}{MultiSet object}
}
\value{
matrix of background intensities
}
\description{
Extracts the matrix of background intensities (bg)
from a MultiSet object created by \link{readArrays}
}

