% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blockScaleNorm-methods.R
\docType{methods}
\name{blockScaleNorm}
\alias{blockScaleNorm}
\alias{blockScaleNorm,MultiSet-method}
\title{Within array scaling normalisation of blocks (print-tip groups)}
\usage{
blockScaleNorm(x, ...)

\S4method{blockScaleNorm}{MultiSet}(x)
}
\arguments{
\item{x}{MultiSet object with fMedian matrices in the assayData slot}

\item{arr}{integer value corresponding to array to be normalised}
}
\value{
MultiSet object with normalised foreground signal
in the fMedian matrix
}
\description{
Within array scaling normalisation of blocks (print-tip groups)
}

