<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AccessorMethods.R
=======
% Generated by roxygen2 (4.0.1): do not edit by hand
>>>>>>> ac19e9f619a67b75113558410c66eb275dc1e1db
\docType{methods}
\name{bmean}
\alias{bmean}
\alias{bmean,MultiSet-method}
\title{Mean Background Intensity Accessor}
\usage{
bmean(x)

\S4method{bmean}{MultiSet}(x)
}
\arguments{
\item{x}{MultiSet object}
}
\value{
matrix of mean background intensities
}
\description{
Extracts the matrix of mean background intensities (bMean) from a MultiSet object created by \link{readArrays}
}

