% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dia}
\alias{dia}
\alias{dia,MultiSet-method}
\title{Diameter Accessor}
\usage{
dia(x)

\S4method{dia}{MultiSet}(x)
}
\arguments{
\item{x}{MultiSet object}
}
\value{
matrix of feature diameters
}
\description{
Extracts the matrix of feature diameter values from a MultiSet object created by \link{readArrays}
}

