\docType{methods}
\name{fg}
\alias{fg}
\alias{fg,MultiSet-method}
\title{Foreground Intensity Accessor}
\usage{
fg(x)

\S4method{fg}{MultiSet}(x)
}
\arguments{
  \item{x}{MultiSet object}
}
\value{
matrix of foreground intensities
}
\description{
Extracts the matrix of foreground intensities (fMedian)
from a MultiSet object created by \link{readArrays}
}

