% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AccessorMethods.R
\docType{methods}
\name{flags}
\alias{flags}
\alias{flags,MultiSet-method}
\title{Flags Accessor}
\usage{
flags(x)

\S4method{flags}{MultiSet}(x)
}
\arguments{
\item{x}{MultiSet object}
}
\value{
matrix of flagged values
}
\description{
Extracts the matrix of flagged values from a MultiSet object created by \link{readArrays}
}

