\docType{methods}
\name{getArrayLayout}
\alias{getArrayLayout}
\title{Array Layout Accessor}
\arguments{
  \item{x}{MultiSet object with block, column and row
  information in the fData slot}
}
\value{
  layout matrix
}
\description{
  Gets microarray layout information matrix Used to obtain
  the layout information for \link{plotImage}
}

