% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maplot-methods.R
\docType{methods}
\name{maplot}
\alias{maplot}
\title{MA-Plot}
\usage{
maplot(x, ...)
}
\arguments{
\item{x}{ExpressionSet object}

\item{...}{other arguments passed to plot}

\item{arr1}{integer corresponding to 1st array to be plotted}

\item{arr2}{integer corresponding to 2nd array to be plotted}
}
\value{
Plot created on current graphics device
}
\description{
Creates simple MA plot
}

