\name{peparray-package}
\alias{peparray-package}
\alias{peparray}
\docType{package}
\title{
Pre-processing and analysis of peptide microarray data.
}
\description{
peparray is a package for the pre-processing and exploratory analysis of spotted peptide microarray data. It includes functions for reading raw array data from Genepix GPR files, quality contol, normalisation and annotation.
}
\details{
\tabular{ll}{
Package: \tab peparray\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-05-20\cr
License: \tab GPL-2\cr
}

}
\author{
Kate Nambiar <k.z.nambiar@bsms.ac.uk>

}
