% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotImage}
\alias{plotImage}
\alias{plotImage,MultiSet-method}
\title{Image plot of microarray data}
\usage{
plotImage(x, ...)

\S4method{plotImage}{MultiSet}(x, arr = 1, slot = "bg", lowcol, highcol,
  ncols = 123, ...)
}
\arguments{
\item{x}{MultiSet object with fMedian and/or bMedian matrices in the assayData slot
and layout data (block, column and row) in the fData slot.}

\item{lowcol}{Colour associated with low signal intensities}

\item{highcol}{Colour associated with high signal intensities}

\item{ncols}{number of colour shades used}
}
\value{
MultiSet object with transformed and background corrected foreground signal in the fMedian matrix
}
\description{
Creates image plot representing the microarray foreground or background signal intensity
organised by the spatial location of the spots on the array.
}

