% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSubarrayClosestValues-methods.R
\docType{methods}
\name{plotSubarrayClosestValues}
\alias{plotSubarrayClosestValues}
\alias{plotSubarrayClosestValues,MultiSet-method}
\title{Scatter plot of closest two values from subarrays of peptide microarray data}
\usage{
plotSubarrayClosestValues(x, ...)

\S4method{plotSubarrayClosestValues}{MultiSet}(x, arr, transform = "log2",
  ...)
}
\arguments{
\item{x}{MultiSet object with fMedian matrix in the assayData slot}

\item{arr}{Index indicating which array should be plotted}

\item{transform}{function to apply to transform the raw data}
}
\value{
plot on current graphics device
}
\description{
Scatter plot of closest two values from subarrays of peptide microarray data
}

