% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotSubarrays-methods.R
\docType{methods}
\name{plotSubarrays}
\alias{plotImage,MultiSet-method}
\alias{plotSubarrays}
\title{Scatter plots of subarrays from peptide microarray data}
\usage{
plotSubarrays(x, ...)

\S4method{plotImage}{MultiSet}(x, arr, subarray = c(1, 2),
  transform = "log2", ...)
}
\arguments{
\item{x}{MultiSet object with fMedian and/or bMedian matrices in the assayData slot}

\item{arr}{Index indicating which array should be plotted}

\item{subarray}{Vector of length = 2 indicating which subarrays should be plotted}

\item{transform}{function to apply to transform the raw data}
}
\value{
plot on current graphics device
}
\description{
Scatter plots of subarrays from peptide microarray data
}

