% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotSubarrayScatter-methods.R
\docType{methods}
\name{plotSubarrayScatter}
\alias{plotSubarrayScatter}
\alias{plotSubarrayScatter,MultiSet-method}
\title{Scatter plots of subarrays from peptide microarray data}
\usage{
plotSubarrayScatter(x, ...)

\S4method{plotSubarrayScatter}{MultiSet}(x, arr, subarray = c(1, 2),
  flagval = -100, transform = "log2", ...)
}
\arguments{
\item{x}{MultiSet object with fMedian matrix in the assayData slot}

\item{arr}{Index indicating which array should be plotted}

\item{subarray}{Vector of length = 2 indicating which subarrays should be plotted}

\item{transform}{function to apply to transform the raw data}
}
\value{
plot on current graphics device
}
\description{
Scatter plots of subarrays from peptide microarray data
}

