% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readArrayHeader}
\alias{readArrayHeader}
\title{Read peptide microarray data from GPR files}
\usage{
readArrayHeader(x, wavelength)
}
\arguments{
\item{x}{filename including full path for the array header to be read}

\item{wavelength}{integer value for the scan wavelength (typically 635 for Cy5 and 532 for Cy3)}
}
\value{
data frame
}
\description{
\code{readArrayHeader} is usually called by \code{readArrays} rather than being called directly by the user.
}

