% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scaleNorm.R
\name{scaleNorm}
\alias{scaleNorm}
\title{Median Scale Normalisation}
\usage{
scaleNorm(x)
}
\arguments{
\item{x}{matrix of intensity values with
probes as rows and samples in columns}
}
\value{
matrix of normalised intensities
}
\description{
Median Scale Normalisation
}

