\name{scaleNormControl}
\alias{scaleNormControl}
\title{Scale Normalisation To Control Probe}
\usage{
scaleNormControl(x, controlID)

scaleNormControl(x, controlID)
}
\arguments{
  \item{x}{matrix of intensity values with probes as rows
  and samples in columns}

  \item{x}{matrix of intensity values with probes as rows
  and samples in columns}
}
\value{
matrix of normalised intensities

matrix of normalised intensities
}
\description{
Scale Normalisation To Control Probe

Scale Normalisation using GMM
}

