% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotateArrays-methods.R
\docType{methods}
\name{annotateArrays}
\alias{annotateArrays}
\alias{annotateArrays,ExpressionSet-method}
\alias{annotateArrays,MultiSet-method}
\title{Adds annotation to peptide microarray data}
\usage{
annotateArrays(x, ...)

\S4method{annotateArrays}{MultiSet}(x, pheno = NULL, protocol = NULL,
  feature = NULL)

\S4method{annotateArrays}{ExpressionSet}(x, pheno = NULL, protocol = NULL,
  feature = NULL)
}
\arguments{
\item{x}{MultiSet object}
}
\value{
Multiset object
}
\description{
Adds annotation to peptide microarray data
}

