\name{arrayAve-methods}
\docType{methods}
\alias{arrayAve}
\alias{arrayAve,pepArrayPP-method}
\title{Methods for function arrayAve}
\description{
Calculates the mean of intra-array replicate features.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "pepArrayPP")}}{
Calculates the mean of the foreground intensity from a pepArrayPP object and produces a new pepArray object with the averaged intensities in the exprs slot
}
}
}

\keyword{methods}
