% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/arraySummary-methods.R
\docType{methods}
\name{arraySummary}
\alias{arraySummary}
\alias{arraySummary,MultiSet-method}
\title{Summarises intra-array replicates with flagged values excluded}
\usage{
arraySummary(x, ...)

\S4method{arraySummary}{MultiSet}(x, method, cv.threshold)
}
\arguments{
\item{x}{MultiSet object}

\item{method}{c("Median", "Mean")}
}
\value{
ExpressionSet with summarised intensity values in exprs slot
}
\description{
Summarises intra-array replicates with flagged values excluded
}

