% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{bmean}
\alias{bmean}
\alias{bmean,MultiSet-method}
\title{Mean Background Intensity Accessor}
\usage{
bmean(x)

\S4method{bmean}{MultiSet}(x)
}
\arguments{
\item{x}{MultiSet object}
}
\value{
matrix of mean background intensities
}
\description{
Extracts the matrix of mean background intensities (bMean)
from a MultiSet object created by \link{readArrays}
}

