% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AccessorMethods.R
\docType{methods}
\name{bmedian}
\alias{bmedian}
\alias{bmedian,MultiSet-method}
\title{Median Background Intensity Accessor}
\usage{
bmedian(x)

\S4method{bmedian}{MultiSet}(x)
}
\arguments{
\item{x}{MultiSet object}
}
\value{
matrix of median background intensities
}
\description{
Extracts the matrix of median background intensities (bMedian) from a MultiSet object created by \link{readArrays}
}

