% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AccessorMethods.R
\docType{methods}
\name{fg}
\alias{fg}
\alias{fg,MultiSet-method}
\title{Foreground Intensity Accessor}
\usage{
fg(x)

\S4method{fg}{MultiSet}(x)
}
\arguments{
\item{x}{MultiSet object}
}
\value{
matrix of foreground intensities
}
\description{
Extracts the matrix of foreground intensities (fMedian) from a MultiSet object created by \link{readArrays}
}

