\docType{methods}
\name{flags}
\alias{flags}
\alias{flags,MultiSet-method}
\title{Flags Accessor}
\usage{
flags(x)

\S4method{flags}{MultiSet}(x)
}
\arguments{
  \item{x}{MultiSet object}
}
\value{
matrix of background intensities
}
\description{
Extracts the matrix of flagged values from a MultiSet
object created by \link{readArrays}
}

