\docType{methods}
\name{maplot}
\alias{maplot}
\title{MA-Plot}
\usage{
maplot(x, ...)
}
\arguments{
  \item{x}{MultiSet or ExpressionSet object}

  \item{arr1}{integer corresponding to 1st array to be
  plotted}

  \item{arr2}{integer corresponding to 2nd array to be
  plotted}

  \item{...}{other arguments passed to plot}
}
\value{
Plot created on current graphics device
}
\description{
Creates simple MA plot
}

