\name{pippa-package}
\alias{pippa-package}
\alias{pippa}
\docType{package}
\title{
Pre-processing and analysis of peptide microarray data.
}
\description{
peparray is a package for the pre-processing and exploratory analysis of spotted peptide microarray data. It includes functions for reading raw array data from Genepix GPR files, quality contol, normalisation and annotation.
}
\details{
\tabular{ll}{
Package: \tab pippa\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-09-21\cr
License: \tab GPL-3\cr
}

}
\author{
Kate Nambiar <k.z.nambiar@bsms.ac.uk>

}
