% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readArrayHeader.R
\name{readArrayHeader}
\alias{readArrayHeader}
\title{Reads header information from GAL or GPR files}
\usage{
readArrayHeader(x, wavelength, wavelength.field = "Wavelengths")
}
\arguments{
\item{x}{filename including full path for the array header to be read}
}
\value{
data frame
}
\description{
\code{readArrayHeader} is usually called by \code{readArrays} rather than being called directly by the user.
}

