% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scaleNorm-methods.R
\docType{methods}
\name{scaleNorm}
\alias{scaleNorm}
\alias{scaleNorm,ExpressionSet-method}
\alias{scaleNorm,MultiSet-method}
\title{Scale Normalisation}
\usage{
scaleNorm(x, ...)

\S4method{scaleNorm}{MultiSet}(x, controlID = NULL)

\S4method{scaleNorm}{ExpressionSet}(x, controlID = NULL)
}
\arguments{
\item{x}{MultiSet object with fMedian matrices in the assayData slot}

\item{offset}{numeric value added to raw signal intensity before background correction is implemented}
}
\value{
MultiSet object with normalised foreground signal in the fMedian matrix
}
\description{
Scale Normalisation
}

