% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scaleNorm.R
\name{scaleNormControl}
\alias{scaleNormControl}
\title{Scale Normalisation To Control Probe}
\usage{
scaleNormControl(x, controlID)
}
\arguments{
\item{x}{matrix of intensity values with
probes as rows and samples in columns}
}
\value{
matrix of normalised intensities
}
\description{
Scale Normalisation To Control Probe
}

